requirements.txt
pyautogen==0.2.34
matplotlib
yfinance
pyautogen[websockets]==0.2.34
fastapi
uvicorn
streamlit